+++ module/system/preloadTemplates.js
@@ -6,9 +6,13 @@ export const preloadHandlebarsTemplates = async function () {
         'systems/chroniclesystem/templates/items/partials/physical-item.html',
         'systems/chroniclesystem/templates/items/partials/equipment-item.html',
 
+        'systems/chroniclesystem/templates/items/tabs/technique-details-tab.html',
+        'systems/chroniclesystem/templates/items/tabs/technique-works-tab.html',
+
         'systems/chroniclesystem/templates/actors/partials/tabs/abilities-tab.html',
         'systems/chroniclesystem/templates/actors/partials/tabs/combat-and-intrigue-tab.html',
         'systems/chroniclesystem/templates/actors/partials/tabs/qualities-tab.html',
+        'systems/chroniclesystem/templates/actors/partials/tabs/sorcery-tab.html',
         'systems/chroniclesystem/templates/actors/partials/tabs/equipments-tab.html',
         'systems/chroniclesystem/templates/actors/partials/tabs/description-tab.html',
         'systems/chroniclesystem/templates/actors/partials/form-group.html',